+++ navi/test/mocks/mocks.dart
@@ -1,3 +1,6 @@
+export 'data/books.dart';
+export 'logging.dart';
 export 'mock_app.dart';
+export 'mock_navi_router_delegate.dart';
 export 'mock_route_information_provider.dart';
 export 'navigation_method.dart';